[git actions] Dameng tests action, dameng connector refactoring, dependencies update, minor structure refactorng
