+++ .github/workflows/scripts/damengStartup.sh
@@ -1,0 +1,33 @@
+#!/bin/bash
+#  create user "onlyoffice" identified by "onlyoffice" password_policy 0;
+/opt/dmdbms/bin/disql SYSDBA/SYSDBA001:5236 <<EOF
+  CREATE TABLE doc_changes
+  (
+  tenant varchar(255) NOT NULL,
+  id varchar(255) NOT NULL,
+  change_id int NOT NULL,
+  user_id varchar(255) NOT NULL,
+  user_id_original varchar(255) NOT NULL,
+  user_name varchar(255) NOT NULL,
+  change_data text NOT NULL,
+  change_date TIMESTAMP(6) NOT NULL,
+  PRIMARY KEY (tenant, id, change_id)
+  );
+  CREATE TABLE task_result
+  (
+  tenant varchar(255) NOT NULL,
+  id varchar(255) NOT NULL,
+  status int NOT NULL,
+  status_info int NOT NULL,
+  created_at TIMESTAMP(6) DEFAULT NOW(),
+  last_open_date TIMESTAMP(6) NOT NULL,
+  user_index int NOT NULL DEFAULT 1,
+  change_id int NOT NULL DEFAULT 0,
+  callback text NOT NULL,
+  baseurl text NOT NULL,
+  password text NULL,
+  additional text NULL,
+  PRIMARY KEY (tenant, id)
+  );
+  exit;
+EOF
