+++ .github/workflows/unitTests.yml
@@ -4,7 +4,7 @@ on:
     paths:
       - '**.js'
       - '!tests/integration/**'
-      - '!DocService/sources/**Connector.js'
+      - '!DocService/sources/databaseConnectors/**'
 jobs:
   unit-tests:
     name: Service unit tests
@@ -13,6 +13,16 @@ jobs:
     steps:
       - name: Check out repository code
         uses: actions/checkout@v3
+
+      - name: Caching dependencies
+        uses: actions/setup-node@v3
+        with:
+          node-version: '18'
+          cache: 'npm'
+          cache-dependency-path: |
+            ./npm-shrinkwrap.json
+            ./Common/npm-shrinkwrap.json
+            ./DocService/npm-shrinkwrap.json
 
       - name: Install modules
         run: |
