+++ Common/config/default.json
@@ -177,7 +177,13 @@
 				"connectionlimit": 10,
 				"max_allowed_packet": 1048575,
 				"pgPoolExtraOptions": {},
-				"damengExtraOptions": {},
+				"damengExtraOptions": {
+					"columnNameUpperCase": false,
+					"columnNameCase": "lower",
+					"loginEncrypt": false,
+					"localTimezone": 0,
+					"compatibleMode": "mysql"
+				},
 				"oracleExtraOptions": {},
 				"msSqlExtraOptions": {
 					"options": {
