+++ DocService/sources/canvasservice.js
@@ -38,7 +38,7 @@ var co = require('co');
 const ms = require('ms');
 const retry = require('retry');
 const MultiRange = require('multi-integer-range').MultiRange;
-var sqlBase = require('./baseConnector');
+var sqlBase = require('./databaseConnectors/baseConnector');
 const utilsDocService = require('./utilsDocService');
 var docsCoServer = require('./DocsCoServer');
 var taskResult = require('./taskresult');
