+++ DocService/sources/changes2forgotten.js
@@ -41,7 +41,7 @@ var utils = require('./../../Common/sources/utils');
 const storage = require('./../../Common/sources/storage-base');
 const queueService = require('./../../Common/sources/taskqueueRabbitMQ');
 const operationContext = require('./../../Common/sources/operationContext');
-const sqlBase = require('./baseConnector');
+const sqlBase = require('./databaseConnectors/baseConnector');
 const docsCoServer = require('./DocsCoServer');
 const taskResult = require('./taskresult');
 const editorDataStorage = require('./' + config.get('services.CoAuthoring.server.editorDataStorage'));
