+++ DocService/sources/databaseConnectors/mssqlConnector.js
@@ -35,7 +35,7 @@
 const sql = require("mssql");
 const config = require('config');
 const connectorUtilities = require('./connectorUtilities');
-const utils = require('./../../Common/sources/utils');
+const utils = require('../../../Common/sources/utils');
 
 const configSql = config.get('services.CoAuthoring.sql');
 const cfgTableResult = configSql.get('tableResult');
@@ -175,7 +175,7 @@ async function executeBulk(ctx, table) {
 }
 
 function closePool() {
-  sql.close();
+  return sql.close();
 }
 
 function addSqlParameterObjectBased(parameter, name, type, accumulatedObject) {
