+++ DocService/sources/databaseConnectors/oracleConnector.js
@@ -35,7 +35,7 @@
 const oracledb = require('oracledb');
 const config = require('config');
 const connectorUtilities = require('./connectorUtilities');
-const utils = require('./../../Common/sources/utils');
+const utils = require('../../../Common/sources/utils');
 
 const configSql = config.get('services.CoAuthoring.sql');
 const cfgTableResult = configSql.get('tableResult');
@@ -145,7 +145,7 @@ async function executeBunch(ctx, sqlCommand, values = [], noLog = false) {
 }
 
 function closePool() {
-  pool?.close(forceClosingCountdownMs);
+  return pool?.close(forceClosingCountdownMs);
 }
 
 function healthCheck(ctx) {
