+++ FileConverter/sources/converter.js
@@ -45,7 +45,7 @@ var commonDefines = require('./../../Common/sources/commondefines');
 var storage = require('./../../Common/sources/storage-base');
 var utils = require('./../../Common/sources/utils');
 var constants = require('./../../Common/sources/constants');
-var baseConnector = require('./../../DocService/sources/baseConnector');
+var baseConnector = require('../../DocService/sources/databaseConnectors/baseConnector');
 const wopiClient = require('./../../DocService/sources/wopiClient');
 const taskResult = require('./../../DocService/sources/taskresult');
 var statsDClient = require('./../../Common/sources/statsdclient');
